+++ cam_sender/Cargo.toml
@@ -6,14 +6,15 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bincode = "1.3.3"
 bytes = "1.2.1"
+clap = { version = "4.0.11", features = ["derive"] }
 env_logger = "0.9.1"
 futures = "0.3.24"
 futures-core = "0.3.24"
+infer_server = { path = "../infer_server" }
 log = "0.4.17"
 reqwest = { version = "0.11.12", features = ["stream", "multipart"] }
 rscam = "0.5.5"
 tokio = { version = "1.21.2", features = ["full"] }
 tokio-util = { version = "0.7.4", features = ["codec"] }
-infer_server = { path = "../infer_server" }
-bincode = "1.3.3"
